pyproject.toml
[project]
name = "starberry"
version = "0.1.0"
description = "Issue replication codebase."
dependencies = [
    "aiosqlite>=0.20.0",
    "litestar[pydantic,sqlalchemy,standard,structlog]>=2.8.3",
    "strawberry-graphql[debug-server]>=0.229.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
